+++ tibco/src/main/java/nl/nn/adapterframework/extensions/tibco/SendTibcoMessage.java
@@ -15,6 +15,9 @@
  */
 package nl.nn.adapterframework.extensions.tibco;
 
+import java.io.IOException;
+import java.net.URL;
+
 import javax.jms.Connection;
 import javax.jms.ConnectionFactory;
 import javax.jms.DeliveryMode;
@@ -25,14 +28,21 @@ import javax.jms.MessageConsumer;
 import javax.jms.MessageProducer;
 import javax.jms.Session;
 import javax.jms.TextMessage;
+import javax.xml.transform.TransformerConfigurationException;
+import javax.xml.transform.TransformerException;
 
+import nl.nn.adapterframework.configuration.ConfigurationException;
 import nl.nn.adapterframework.core.IPipeLineSession;
 import nl.nn.adapterframework.core.ParameterException;
 import nl.nn.adapterframework.core.PipeRunException;
+import nl.nn.adapterframework.parameters.Parameter;
 import nl.nn.adapterframework.parameters.ParameterResolutionContext;
 import nl.nn.adapterframework.parameters.ParameterValueList;
 import nl.nn.adapterframework.pipes.TimeoutGuardPipe;
+import nl.nn.adapterframework.util.ClassUtils;
 import nl.nn.adapterframework.util.CredentialFactory;
+import nl.nn.adapterframework.util.DomBuilderException;
+import nl.nn.adapterframework.util.TransformerPool;
 
 import org.apache.commons.lang.StringUtils;
 
@@ -79,7 +89,7 @@ import com.tibco.tibjms.admin.TibjmsAdminException;
  * </p>
  * 
  * @author Peter Leeuwenburgh
- * @version $Id$
+ * @version $Id: SendTibcoMessage.java,v 1.9 2016/12/21 10:58:06 m99f706 Exp $
  */
 
 public class SendTibcoMessage extends TimeoutGuardPipe {
@@ -156,8 +166,18 @@ public class SendTibcoMessage extends TimeoutGuardPipe {
 			replyTimeout_work = Integer.parseInt(replyTimeout_work_str);
 		}
 		soapAction_work = getParameterValue(pvl, "soapAction");
-		if (soapAction_work == null) {
+		if (soapAction_work == null)
 			soapAction_work = getSoapAction();
+
+		if (StringUtils.isEmpty(soapAction_work)) {
+			log.debug(getLogPrefix(session) + "deriving default soapAction");
+			try {
+				URL resource = ClassUtils.getResourceURL(this, "/xml/xsl/esb/soapAction.xsl");
+				TransformerPool tp = new TransformerPool(resource, true);
+				soapAction_work = tp.transform(input.toString(), null);
+			} catch (Exception e) {
+				log.error(getLogPrefix(session) + "failed to execute soapAction.xsl");
+			}
 		}
 
 		if (messageProtocol_work == null) {
@@ -222,16 +242,6 @@ public class SendTibcoMessage extends TimeoutGuardPipe {
 			} else {
 				msg.setJMSDeliveryMode(DeliveryMode.PERSISTENT);
 				msgProducer.setDeliveryMode(DeliveryMode.PERSISTENT);
-			}
-			if (StringUtils.isEmpty(soapAction_work)) {
-				if (queueName_work.contains(".")) {
-					String[] q = StringUtils.split(queueName_work, ".");
-					if (q[0].equalsIgnoreCase("P2P") && q.length>=4) {
-						soapAction_work = q[3];
-					} else if (q[0].equalsIgnoreCase("ESB") && q.length>=8) {
-						soapAction_work = q[6] + "_" + q[7];
-					}
-				}
 			}
 			if (StringUtils.isNotEmpty(soapAction_work)) {
 				log.debug(getLogPrefix(session) + "setting [SoapAction] property to value [" + soapAction_work + "]");